+++ Software/i2c.c
@@ -2,8 +2,8 @@
  * Copyright (c) 2025, Dewayne L. Hafenstein.  All rights reserved.
  *
  * Bit-banging I2C implementation for PIC18F27Q43
- * Uses RC3 (SCL) and RC4 (SDA) with open drain configuration   
- * 
+ * Uses RC3 (SCL) and RC4 (SDA) with open drain configuration
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -13,7 +13,7 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- * 
+ *
  */
 #include <xc.h>
 #include "i2c.h"
@@ -64,7 +64,7 @@ uint8_t i2cWriteBuffer(uint8_t address, uint8_t *data, uint8_t length)
     _i2cStart();
 
     // Send address with write bit (0)
-    result = _i2cWriteByte((address << 1) | 0x00);
+    result = _i2cWriteByte((uint8_t)((address << 1) | 0x00));
     if (result != I2C_SUCCESS)
     {
         _i2cStop();
@@ -89,7 +89,7 @@ uint8_t i2cWriteBuffer(uint8_t address, uint8_t *data, uint8_t length)
 
 /****************************************************************************
  * Function: i2cWriteRegister
- * Description: Write a single register on an I2C device.  
+ * Description: Write a single register on an I2C device.
  * Parameters:
  *   address - The I2C device address
  *   reg - The register address to write to
@@ -123,7 +123,7 @@ uint8_t i2cReadBuffer(uint8_t address, uint8_t *data, uint8_t length)
     _i2cStart();
 
     // Send address with read bit (1)
-    result = _i2cWriteByte((address << 1) | 0x01);
+    result = _i2cWriteByte((uint8_t)((address << 1) | 0x01));
     if (result != I2C_SUCCESS)
     {
         _i2cStop();
@@ -159,7 +159,7 @@ uint8_t i2cReadRegister(uint8_t address, uint8_t reg, uint8_t *data)
     _i2cStart();
 
     // Send address with write bit to write register address
-    result = _i2cWriteByte((address << 1) | 0x00);
+    result = _i2cWriteByte((uint8_t)((address << 1) | 0x00));
     if (result != I2C_SUCCESS)
     {
         _i2cStop();
@@ -178,7 +178,7 @@ uint8_t i2cReadRegister(uint8_t address, uint8_t reg, uint8_t *data)
     _i2cRestart();
 
     // Send address with read bit
-    result = _i2cWriteByte((address << 1) | 0x01);
+    result = _i2cWriteByte((uint8_t)((address << 1) | 0x01));
     if (result != I2C_SUCCESS)
     {
         _i2cStop();
@@ -215,33 +215,33 @@ void i2cResetBus(void)
 
 void _i2cSclHigh(void)
 {
-    TRISC |= SCL;       // Ensure SCL is input
+    TRISC |= SCL; // Ensure SCL is input
     I2C_HALF_DELAY();
 }
 
 void _i2cSclLow(void)
 {
-    LATC &= ~SCL;       // Output low
-    TRISC &= ~SCL;      // Set as output
+    LATC &= ~SCL;  // Output low
+    TRISC &= ~SCL; // Set as output
     I2C_HALF_DELAY();
 }
 
 void _i2cSdaHigh(void)
 {
-    TRISC |= SDA;       // Ensure SDA is input
+    TRISC |= SDA; // Ensure SDA is input
     I2C_HALF_DELAY();
 }
 
 void _i2cSdaLow(void)
 {
-    LATC &= ~SDA;        // Output low
-    TRISC &= ~SDA;       // Set as output
+    LATC &= ~SDA;  // Output low
+    TRISC &= ~SDA; // Set as output
     I2C_HALF_DELAY();
 }
 
 uint8_t _i2cSdaRead(void)
 {
-    TRISC |= SDA;       // Ensure SDA is input
+    TRISC |= SDA; // Ensure SDA is input
     I2C_HALF_DELAY();
     return PORTC & SDA; // Read SDA state
 }
